# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "ATWT API for Testing",
        "description": "This is an example api for testing purpose. automatedtestingwithtuyen.com",
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "contact": {
            "name": "Tuyen Nguyen",
            "url": "http://automatedtestingwithtuyen.com",
            "email": "trongtuyen96@gmail.com"
        }
    },
    "host": "atwt-api.herokuapp.com",
    "basePath": "/v1",
...
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/zoo/{zooId}/animals": {
            "get": {
                "tags": [
                    "zoo"
                ],
                "summary": "Find all animals in the zoo by Id",
                "description": "Returns animals in the zoo",
                "operationId": "getAllAnimalsOfZooById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "zooId",
                        "in": "path",
                        "description": "Id of zoo to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "$ref": "#/definitions/Animal"
                        }
                    },
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Zoo not found"
                    }
                }
            }
        },
        "/animal": {
            "get": {
                "tags": [
                    "animal"
                ],
                "summary": "Get animals with conditions",
                "description": "",
                "operationId": "getAnimals",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "name",
                        "description": "Name of animal",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "species",
                        "description": "Species of animal",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "color",
                        "description": "Color of animal",
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "weight",
                        "description": "Weight of animal (greater than or equal)",
                        "type": "integer"
                    },
                    {
                        "in": "query",
                        "name": "lifespan",
                        "description": "Lifespan of animal (greater than or equal)",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "schema": {
                            "type": "array",
                            "$ref": "#/definitions/Animal"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "animal"
                ],
                "summary": "Add a new animal",
                "description": "",
                "operationId": "postAnimal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Animal object that need to add",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Animal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/animal/{animalId}": {
            "get": {
                "tags": [
                    "animal"
                ],
                "summary": "Find animal by Id",
                "description": "Returns a single animal",
                "operationId": "getAnimalById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "animalId",
                        "in": "path",
                        "description": "Id of animal to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Animal"
                        }
                    },
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Animal not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "animal"
                ],
                "summary": "Update an existing animal",
                "description": "",
                "operationId": "putAnimal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "animalId",
                        "in": "path",
                        "description": "Id of animal to update",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Animal object that needs to be updated",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Animal"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Update successfully"
                    },
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Animal not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "animal"
                ],
                "summary": "Delete a animal",
                "description": "",
                "operationId": "deleteAnimal",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "animalId",
                        "in": "path",
                        "description": "Animal Id to delete",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid Id supplied"
                    },
                    "404": {
                        "description": "Animal not found"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Animal": {
            "type": "object",
            "required": [
                "zooID"
            ],
            "properties": {
                "species": {
                    "type": "string"
                },
                "lifespan": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                },
                "zooID": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "Animal"
            }
        }
    }
}